+++ src/utils/case_setup.jl
@@ -55,6 +55,8 @@ end
     _reset_field(x_rh, res_type::ElementReset, 𝒰::UpdateCase, opers::Vector{<:TS.TimePeriod})
     _reset_field(x_rh, res_type::Union{InitReset, TimeWeightReset}, 𝒰::UpdateCase, opers::Vector{<:TS.TimePeriod})
     _reset_field(x_rh, res_type::OperReset, 𝒰::UpdateCase, opers::Vector{<:TS.TimePeriod})
+    _reset_field(x_rh, res_type::PartitionReset, 𝒰::UpdateCase, opers::Vector{<:TS.TimePeriod})
+    _reset_field(x_rh, res_type::EmptyReset, 𝒰::UpdateCase, opers::Vector{<:TS.TimePeriod})
 
 Resets the field expressed through `res_type` of element `x_rh` with the new value. The type
 of the new value is depending on the specified `res_type`:
@@ -63,6 +65,10 @@ of the new value is depending on the specified `res_type`:
 2. `res_type::Union{InitReset, TimeWeightReset}` uses the value in `res_type` directly,
 3. `res_type::OperReset` creates a new operational profile based on the original
    operational profile in `res_type` and the set of operational periods `opers`.
+4. `res_type::PartitionReset` creates a new partition profile based on the original
+   partition profile in `res_type` and the set of operational periods `opers`.
+5. `res_type::EmptyReset` does not reset any field and is used to avoid problems with
+   partition profile resetting.
 """
 function _reset_field(
     x_rh,
@@ -91,10 +97,37 @@ function _reset_field(
     @reset res_type.lens(x_rh) = OperationalProfile(res_type.val[opers])
     return x_rh
 end
+function _reset_field(
+    x_rh,
+    res_type::PartitionReset,
+    𝒰::UpdateCase,
+    opers::Vector{<:TS.TimePeriod},
+)
+    # Extract the required variables from the UpdateCase
+    𝒯 = get_time_struct(𝒰)
+
+    # Identify the partitions of the original problem that are used within the current
+    # receding horizon problem
+    𝒯ᵖᵈ = period_duration(res_type, 𝒯)
+    parts = filter(t_pd -> isempty(setdiff(t_pd, opers)), 𝒯ᵖᵈ)
+
+    # Reset the partition profile of the receding horizon problem based on the relevant
+    # partitions
+    @reset res_type.lens(x_rh) = PartitionProfile(res_type.val[parts])
+    return x_rh
+end
+function _reset_field(
+    x_rh,
+    res_type::EmptyReset,
+    𝒰::UpdateCase,
+    opers::Vector{<:TS.TimePeriod},
+)
+    return x_rh
+end
 
 """
-    _create_updatetype(modeltype::RecHorEnergyModel)
-    _create_updatetype(modeltype::RecHorEnergyModel, case::AbstractCase)
+    _create_updatetype(case::AbstractCase, modeltype::RecHorEnergyModel)
+    _create_updatetype(𝒯::TS.TimeStructure, modeltype::RecHorEnergyModel)
 
 Initialize an [`UpdateCase`](@ref) based on the provided [`RecHorEnergyModel`](@ref)
 `modeltype`.
@@ -102,20 +135,27 @@ Initialize an [`UpdateCase`](@ref) based on the provided [`RecHorEnergyModel`](@
 Initialize and populate the [`UpdateCase`](@ref) if the function has as first argument an
 [`AbstractCase`](@extref EnergyModelsBase.AbstractCase).
 """
-function _create_updatetype(modeltype::RecHorEnergyModel)
-    paths_model = _find_update_paths(modeltype)
-    reset_model = AbstractReset[ResetType(field_id, field_id[end], modeltype) for field_id ∈ paths_model]
-    𝒰 = UpdateCase(Substitution(modeltype, reset_model), Dict(), Dict(), ProductSub[], Vector[])
-    _init_mapping!(𝒰, modeltype)
-    return 𝒰
-end
 function _create_updatetype(case::AbstractCase, modeltype::RecHorEnergyModel)
     # Create the `UpdateCase` based on the original `Case` structure
-    𝒰 = _create_updatetype(modeltype)
+    𝒰 = _create_updatetype(get_time_struct(case), modeltype)
     _add_elements!(𝒰, get_products(case))
     for 𝒳 ∈ get_elements_vec(case)
         _add_elements!(𝒰, 𝒳)
     end
+    return 𝒰
+end
+function _create_updatetype(𝒯::TS.TimeStructure, modeltype::RecHorEnergyModel)
+    paths_model = _find_update_paths(modeltype)
+    reset_model = AbstractReset[ResetType(field_id, field_id[end], modeltype) for field_id ∈ paths_model]
+    𝒰 = UpdateCase(
+        𝒯,
+        Substitution(modeltype, reset_model),
+        Dict(),
+        Dict(),
+        ProductSub[],
+        Vector[],
+    )
+    _init_mapping!(𝒰, modeltype)
     return 𝒰
 end
 
@@ -219,3 +259,129 @@ _type_to_key(::Type{T}) where {T<:Union{EMB.Node, NodeSub}} = :nodes
 _type_to_key(::Type{T}) where {T<:Union{Link, LinkSub}} = :links
 _type_to_key(::Type{T}) where {T<:Union{FutureValue, FutureValueSub}} = :future_values
 _type_to_key(::Type{T}) where {T<:Union{EnergyModel, ModelSub}} = :modeltype
+
+"""
+    _check_period_partitions(𝒰::UpdateCase, ℋ::AbstractHorizons, optimizer)
+    _check_period_partitions(log::Dict{String, Vector{String}}, 𝒮::Vector{<:AbstractSub}, 𝒯::TS.TimeStructure, ℋ::AbstractHorizons, optimizer)
+    _check_period_partitions(log::Dict{String, Vector{String}}, s::AbstractSub, 𝒯::TS.TimeStructure, ℋ::AbstractHorizons, optimizer)
+
+Function for checking that the individual period partitions are in line with the horizon
+structure `ℋ`.
+
+Returns an `AssertionError` if any period partition is not consistent with `ℋ`. It furthermore
+creates a log which points towards any inconsistent partitions and horizons.
+"""
+function _check_period_partitions(𝒰::UpdateCase, ℋ::AbstractHorizons, optimizer)
+    # Extract the information from the `UpdateCase`
+    𝒯 = get_time_struct(𝒰)
+
+    # Collect all potential problems
+    log = Dict{String, Vector{String}}()
+    _check_period_partitions(log, get_sub_model(𝒰), 𝒯, ℋ, optimizer)
+    _check_period_partitions(log, get_sub_products(𝒰), 𝒯, ℋ, optimizer)
+    for 𝒮 ∈ get_sub_elements_vec(𝒰)
+        _check_period_partitions(log, 𝒮, 𝒯, ℋ, optimizer)
+    end
+
+    # Throw an `ErrorException` if there are inconsistent partitions
+    !isempty(log) && EMB.compile_logs(𝒰, log)
+end
+
+function _check_period_partitions(
+    log::Dict{String, Vector{String}},
+    𝒮::Vector{<:AbstractSub},
+    𝒯::TS.TimeStructure,
+    ℋ::AbstractHorizons,
+    optimizer,
+)
+    for s ∈ 𝒮
+        _check_period_partitions(log, s, 𝒯, ℋ, optimizer)
+    end
+end
+function _check_period_partitions(
+    log::Dict{String, Vector{String}},
+    s::AbstractSub,
+    𝒯::TS.TimeStructure,
+    ℋ::AbstractHorizons,
+    optimizer,
+)
+    opers = collect(𝒯)
+
+    # Identify whether the `Substitution` has `PartitionReset`s
+    part_res = filter(res_type -> isa(res_type, PartitionReset), resets(s))
+
+    if !isempty(part_res)
+        # Extract the period duration for the reset. It must be the same for all resets
+        𝒯ᵖᵈ = period_duration(first(part_res), 𝒯)
+        opers = collect(𝒯)
+
+        # Iterate through all horizons
+        x_org = original(s)
+        opers_opt_ref = nothing
+        n_opt_ref = nothing
+        n_impl_ref = nothing
+        log["$(x_org)"] = String[]
+        for 𝒽 ∈ ℋ
+            # Extract variables and reassign values
+            opers_opt = opers[indices_optimization(𝒽)]
+            opers_impl = opers[indices_implementation(𝒽)]
+            if isnothing(opers_opt_ref)
+                opers_opt_ref = deepcopy(opers_opt)
+            end
+            if length(opers_opt) < length(opers_opt_ref)
+                n_opt_ref = nothing
+                n_impl_ref = nothing
+            end
+
+            # Check the two horizons
+            msg_opt, bool_opt, n_opt_ref =
+                _check_horizon(n_opt_ref, 𝒯ᵖᵈ, opers_opt, "optimization", optimizer)
+            msg_impl, bool_impl, n_impl_ref =
+                _check_horizon(n_impl_ref, 𝒯ᵖᵈ, opers_impl, "implementation", optimizer)
+
+            # Add to the logs
+            (bool_opt || bool_impl) && push!(log["$(x_org)"], "Horizon $(𝒽): \n")
+            if bool_opt
+                log["$(x_org)"][end] *= msg_opt
+            end
+            if bool_impl
+                log["$(x_org)"][end] *= msg_impl
+            end
+        end
+    end
+end
+
+"""
+    _check_horizon(n_ref::Union{Nothing, Vector{Int}}, 𝒯ᵖᵈ, opersᵣₕ, name::String, optimizer)
+
+Function for identifying any potential inconsistencies between the operational periods of a
+horizon `opersᵣₕ` and the period partitions `𝒯ᵖᵈ` for a given horizon `name`.
+
+Returns a log, a boolean indicating whether there are inconsistencies and the reference
+number of partitions within the horizon.
+"""
+function _check_horizon(n_ref::Union{Nothing, Vector{Int}}, 𝒯ᵖᵈ, opersᵣₕ, name::String, optimizer)
+    parts = filter(t_pd -> isempty(setdiff(t_pd, opersᵣₕ)), 𝒯ᵖᵈ)
+    n_parts = [length(part) for part ∈ parts]
+    log = ""
+    bool_op = nothing
+    bool_poi = false
+    try
+        opersₚₐᵣₜ = reduce(vcat, [collect(part) for part ∈ parts])
+        bool_op = any(t ∉ opersₚₐᵣₜ for t ∈ opersᵣₕ)
+        if isnothing(n_ref)
+            n_ref = n_parts
+        else
+            log = _check_number_op(n_parts, n_ref, name, optimizer)
+            bool_poi = !isempty(log)
+        end
+    catch e
+        bool_op = true
+    end
+    if bool_op
+        log *= "  The partitions are inconsistent with the $(name) horizon.\n"
+    end
+    return log, bool_op || bool_poi, n_ref
+end
+
+_check_number_op(_::Vector{Int}, _::Vector{Int}, _::String, _) = ""